cloud_function_src/requirements.txt
functions-framework==3.*
google-cloud-pubsub>=2.13.0
cloudevents==1.*